+++ Merchant.Domain/ViewModels/CheckoutViewModel.cs
@@ -14,15 +14,15 @@ namespace Merchant.Domain.ViewModels
             CurrencyCode = "GBP";
             CheckoutProductViewModels = new List<CheckoutProductViewModel>()
             {
-                new CheckoutProductViewModel{ Title="Classic White T Shirt Size M", Price=24.99, Delivery=0.00},
-                new CheckoutProductViewModel{ Title="Classic Oxford Shirt Size M", Price=24.99, Delivery=0.00}
+                new CheckoutProductViewModel{ Title="Classic White T Shirt Size M", Price=24.99m, Delivery=0.00m},
+                new CheckoutProductViewModel{ Title="Classic Oxford Shirt Size M", Price=24.99m, Delivery=0.00m}
             };
         }
 
         public IEnumerable<CheckoutProductViewModel> CheckoutProductViewModels;
         public string SellerName { get; set; }
         public string SellerMerchantId { get; }
-        public double TotalCost => CheckoutProductViewModels.Sum(x => x.Price + x.Delivery);
+        public decimal TotalCost => CheckoutProductViewModels.Sum(x => x.Price + x.Delivery);
         public string CurrencyCode { get; }
 
         public string UserId { get; }
@@ -31,7 +31,7 @@ namespace Merchant.Domain.ViewModels
     public class CheckoutProductViewModel
     {
         public string Title { get; set; }
-        public double Price { get; set; }
-        public double Delivery { get; set; }
+        public decimal Price { get; set; }
+        public decimal Delivery { get; set; }
     }
 }
